Move join fixtures to a dedicated module
